\name{dates}
\alias{dates}
\alias{dates<-}
\title{Extract dates from object or assign dates to object.}
\description{Extract dates from objects.  Particularly time series objects.}
\usage{
dates(x, ...)
dates(x) <- value
}
\arguments{
  \item{x}{ an object which contains dates }
  \item{value}{ new dates to be assigned to object }
  \item{\dots}{ further arguments to function }
}
\value{
  a vector of dates
}
\author{ Whit Armstrong }
\keyword{ts}
